Automatically Add's free-gift to cart
This script runs when the checkbox auto_add_gift setting is set to true it simulates a click on the free gift button to add it without a user needing to click the button. This is not how I would like to achieve the automatic adding I would like to call the function which runs when the button is clicked without the need of simulating a click.